Modified to filter acceleration with floating point

+++ lib/MPU6050/MPU6050.hpp
@@ -5,7 +5,7 @@
 #include <vector.hpp>
 #include <filter.hpp>
 
-extern vect_t<int16_t> accel_filtered;
+extern vect_t<float> accel_filtered;
 
 class MPU6050{
 private:
@@ -17,9 +17,9 @@ private:
     static vect_t<int16_t> accel;
     static vect_t<int16_t> gyro;
 
-    static Filter<int16_t, 2> filter_x;
-    static Filter<int16_t, 2> filter_y;
-    static Filter<int16_t, 2> filter_z;
+    static Filter<float, 2> filter_x;
+    static Filter<float, 2> filter_y;
+    static Filter<float, 2> filter_z;
 
     static bool connected;
